add Orb challenge (#9)
Why:
As new features are released we want to keep this project updated, this PR takes care of adding a challenge that covers the concept of using Orbs. It requires the trainee to set the Orb and invoke it with a parameter to get a passing build.

What:
This adds an example config.yml file missing a few key items, and is currently failing. It also adds a README.md that contains information about the challenge.