Add cache challenge (#10)
Why:
As new features are released we want to keep this project updated, this PR takes care of adding a challenge that covers the concept of using Caching. It requires the trainee to update the config to save and restore a cache. Then push a build that saves it, a build that uses the cache, and one that 'busts' the cache.

What:
This adds an example config.yml file with some steps commented out -- they need to be uncommented and updated. It also adds a README.md that contains information about the challenge and a package.json file to be used in caching.